rename: constants.js
* contants -> constants 오타 수정
